+++ uefi-raw/src/protocol/mod.rs
@@ -16,4 +16,5 @@ pub mod memory_protection;
 pub mod misc;
 pub mod network;
 pub mod rng;
+pub mod scsi;
 pub mod shell_params;